+++ src/supabase/supabaseClient.d.ts
@@ -8,7 +8,7 @@ const supabaseUrl = process.env.SUPABASE_URL;
 const supabaseServiceRole = process.env.SUPABASE_SERVICE_ROLE_KEY;
 
 if (!supabaseUrl) {
-    throw new Error("Supabase URL  is missing");
+    throw new Error("Supabase URL is missing");
 }
 if (!supabaseServiceRole) {
     throw new Error("Supabase AnonKey is missing");
